Version 0.1 for Argonaut 6.1.x and Spray 1.3.x.
Based largely on PlayJsonSupport and argonaut-spray. I had a bunch of problems with the latter due to overly excited implicit conversions (resulting in StackOverflow) and the general state of the project (sans tests and structure) led me to create something fresh and dedicated (as it was really just a port of PlayJsonSupport).

I've tried not to use anything overly fancy from either project so future maintainers can simply amend.